Make a proper facade
